requirements.txt
atlassian-jwt==1.8.1
requests-jwt==0.5.3
Django==1.11.15
django-atlassian==0.1.4
django-auth-ldap==1.7.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-debug-toolbar==1.11
django-countries==5.5
django-ldapdb==0.9.0
django-pkgconf==0.3.0
django-widget-tweaks==1.4.5
netaddr==0.7.19
pyldap==2.4.37
python-dateutil==2.7.3
python-ldap==3.1.0
pytz==2018.5
six==1.11.0
psycopg2==2.7.5
ptvsd == 4.3.2
redis==2.10.6
requests==2.18.4
PyJWT==1.6.4
ipdb==0.11
git+https://github.com/fluendo/jira.git@OPE-1085
gunicorn==19.9.0
